+++ Modules/Graphics/Direct3D11/src/aurora/BaseBuffer.h
@@ -8,12 +8,12 @@ namespace aurora::modules::graphics::d3d11 {
 		BaseBuffer(UINT bufferType);
 		virtual ~BaseBuffer();
 
-		bool AE_CALL create(Graphics& graphics, uint32_t size, Usage resUsage, const void* data = nullptr, uint32_t dataSize = 0);
+		bool AE_CALL create(Graphics& graphics, size_t size, Usage resUsage, const void* data = nullptr, size_t dataSize = 0);
 		Usage AE_CALL map(Graphics& graphics, Usage expectMapUsage);
 		void AE_CALL unmap(Graphics& graphics);
-		uint32_t AE_CALL read(uint32_t offset, void* dst, uint32_t dstLen);
-		uint32_t AE_CALL write(Graphics& graphics, uint32_t offset, const void* data, uint32_t length);
-		uint32_t AE_CALL update(Graphics& graphics, uint32_t offset, const void* data, uint32_t length);
+		size_t AE_CALL read(size_t offset, void* dst, size_t dstLen);
+		size_t AE_CALL write(Graphics& graphics, size_t offset, const void* data, size_t length);
+		size_t AE_CALL update(Graphics& graphics, size_t offset, const void* data, size_t length);
 
 		Usage mapUsage;
 		D3D11_MAPPED_SUBRESOURCE mappedRes;
