+++ Modules/Graphics/Direct3D11/src/aurora/ConstantBuffer.h
@@ -12,14 +12,14 @@ namespace aurora::modules::graphics::d3d11 {
 
 		virtual bool AE_CALL isCreated() const override;
 		virtual const void* AE_CALL getNative() const override;
-		virtual bool AE_CALL create(uint32_t size, Usage bufferUsage, const void* data = nullptr, uint32_t dataSize = 0) override;
-		virtual uint32_t AE_CALL getSize() const override;
+		virtual bool AE_CALL create(size_t size, Usage bufferUsage, const void* data = nullptr, size_t dataSize = 0) override;
+		virtual size_t AE_CALL getSize() const override;
 		virtual Usage AE_CALL getUsage() const override;
 		virtual Usage AE_CALL map(Usage expectMapUsage) override;
 		virtual void AE_CALL unmap() override;
-		virtual uint32_t AE_CALL read(uint32_t offset, void* dst, uint32_t dstLen) override;
-		virtual uint32_t AE_CALL write(uint32_t offset, const void* data, uint32_t length) override;
-		virtual uint32_t AE_CALL update(uint32_t offset, const void* data, uint32_t length) override;
+		virtual size_t AE_CALL read(size_t offset, void* dst, size_t dstLen) override;
+		virtual size_t AE_CALL write(size_t offset, const void* data, size_t length) override;
+		virtual size_t AE_CALL update(size_t offset, const void* data, size_t length) override;
 		//virtual void AE_CALL flush() override;
 		virtual bool AE_CALL isSyncing() const override;
 		virtual void AE_CALL destroy() override;
