+++ Modules/Graphics/OpenGL/src/aurora/IndexBuffer.cpp
@@ -20,13 +20,13 @@ namespace aurora::modules::graphics::gl {
 		return this;
 	}
 
-	bool IndexBuffer::create(uint32_t size, Usage bufferUsage, const void* data, uint32_t dataSize) {
+	bool IndexBuffer::create(size_t size, Usage bufferUsage, const void* data, size_t dataSize) {
 		auto rst = _baseBuffer.create(*_graphics.get<Graphics>(), size, bufferUsage, data);
 		_calcNumElements();
 		return rst;
 	}
 
-	uint32_t IndexBuffer::getSize() const {
+	size_t IndexBuffer::getSize() const {
 		return _baseBuffer.size;
 	}
 
@@ -42,15 +42,15 @@ namespace aurora::modules::graphics::gl {
 		_baseBuffer.unmap();
 	}
 
-	uint32_t IndexBuffer::read(uint32_t offset, void* dst, uint32_t dstLen) {
+	size_t IndexBuffer::read(size_t offset, void* dst, size_t dstLen) {
 		return _baseBuffer.read(offset, dst, dstLen);
 	}
 
-	uint32_t IndexBuffer::write(uint32_t offset, const void* data, uint32_t length) {
+	size_t IndexBuffer::write(size_t offset, const void* data, size_t length) {
 		return _baseBuffer.write(offset, data, length);
 	}
 
-	uint32_t IndexBuffer::update(uint32_t offset, const void* data, uint32_t length) {
+	size_t IndexBuffer::update(size_t offset, const void* data, size_t length) {
 		return _baseBuffer.update(offset, data, length);
 	}
