+++ Modules/Graphics/OpenGL/src/aurora/PixelBuffer.cpp
@@ -17,7 +17,7 @@ namespace aurora::modules::graphics::gl {
 		return this;
 	}
 
-	bool PixelBuffer::create(uint32_t size, Usage bufferUsage, const void* data, uint32_t dataSize) {
+	bool PixelBuffer::create(size_t size, Usage bufferUsage, const void* data, size_t dataSize) {
 		destroy();
 
 		bool read = (bufferUsage & Usage::MAP_READ) == Usage::MAP_READ;
@@ -39,7 +39,7 @@ namespace aurora::modules::graphics::gl {
 		}
 	}
 
-	uint32_t PixelBuffer::getSize() const {
+	size_t PixelBuffer::getSize() const {
 		return _baseBuffer.size;
 	}
 
@@ -55,15 +55,15 @@ namespace aurora::modules::graphics::gl {
 		_baseBuffer.unmap();
 	}
 
-	uint32_t PixelBuffer::read(uint32_t offset, void* dst, uint32_t dstLen) {
+	size_t PixelBuffer::read(size_t offset, void* dst, size_t dstLen) {
 		return _baseBuffer.read(offset, dst, dstLen);
 	}
 
-	uint32_t PixelBuffer::write(uint32_t offset, const void* data, uint32_t length) {
+	size_t PixelBuffer::write(size_t offset, const void* data, size_t length) {
 		return _baseBuffer.write(offset, data, length);
 	}
 
-	uint32_t PixelBuffer::update(uint32_t offset, const void* data, uint32_t length) {
+	size_t PixelBuffer::update(size_t offset, const void* data, size_t length) {
 		return _baseBuffer.update(offset, data, length);
 	}
